added permissions

+++ app/Http/Controllers/Admin/AccountController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Account;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AccountController extends Controller
 {
@@ -13,6 +14,9 @@ class AccountController extends Controller
      */
     public function index(Request $request)
     {
+
+        if (!Gate::allows('index account')) abort(404);
+
         $accounts = Account::all();
 
         if($request->expectsJson()) {
@@ -26,11 +30,16 @@ class AccountController extends Controller
 
     public function store(Request $request) {
 
+        if (!Gate::allows('store account')) abort(404);
+
     }
 
 
     public function show(Request $request, Account $account)
     {
+
+        if (!Gate::allows('show product')) abort(404);
+
         if($request->expectsJson()) {
             return response()->json($account);
         }
@@ -41,6 +50,8 @@ class AccountController extends Controller
 
     public function update(Request $request, Account $account) {
 
+        if (!Gate::allows('update product')) abort(404);
+
         if($request->expectsJson()) {
 
         }
@@ -50,6 +61,9 @@ class AccountController extends Controller
 
     public function destroy(Request $request, Account $account)
     {
+
+        if (!Gate::allows('delete product')) abort(404);
+
         $account->delete();
 
         if($request->expectsJson()) {
