+++ app/Http/Controllers/Admin/OrderController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Order;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class OrderController extends Controller
 {
@@ -13,6 +14,8 @@ class OrderController extends Controller
      */
     public function index(Request $request)
     {
+        if (!Gate::allows('index order')) abort(404);
+
         $orders = Order::all();
 
         if($request->expectsJson()) {
@@ -25,6 +28,9 @@ class OrderController extends Controller
 
     public function show(Request $request, Order $order)
     {
+
+        if (!Gate::allows('show order')) abort(404);
+
         if($request->expectsJson()) {
             return response()->json($order);
         }
@@ -35,6 +41,9 @@ class OrderController extends Controller
 
     public function destroy(Request $request, Order $order)
     {
+
+        if (!Gate::allows('delete order')) abort(404);
+
         $order->delete();
 
         if($request->expectsJson()) {
